Add support for BCC/Classic Era client exports
The upcoming 2.5.2 and 1.14.0 clients for BCC and Classic Era have
merged together their interfaces somewhat and have a slightly altered
directory structure as a result for code exports.

Previously all code would have been rooted under the "Interface/" tree,
however with these new patches there are now three directories that
need exporting:

  - Interface/
  - Interface_TBC/
  - Interface_Vanilla/

The ingame "exportinterfacefiles code" command exports all of these
directories as of the most recent public build, including most notably
an export of the "Vanilla" tree if run on the BCC client, and vice-versa
for the "TBC" tree on Classic Era.

For the export tool to correctly export all the interface files going
forward the following changes have been made:

  - Manifest filtering checks for paths matching either "^Interface/" or
    "^Interface_".

  - For filtered exports ("code" or "art") the shortening of paths to
    remove the top-level "Interface" directory has been removed.

The path shortening change is required since, if we attempted to
merge the three interface directories, we'd run into issues with
duplicate filenames - for example "SharedXML/Utils.lua" exists in both
the "Interface_TBC" and "Interface_Vanilla" trees with different
contents.