Hack in WDC4 "support"
The luadbc library hasn't been updated for WDC4 support and both patch 10.1.0
and the upcoming 3.4.2 PTR are using the new database format.

For the use cases required by the interface export tool, we can work around 
the lack of proper WDC4 support by just altering the magic header to claim
the file is WDC3 instead.

This is of course a hack, but in practice it works well enough for now and
allows automated extraction of the interface to continue to function. Once
luadbc is updated (if ever) the hack can be cleaned up with it.