Update casc/dbc libraries to latest versions
Patch 10.1.7 brings about breaking changes to the root manifest that's
processed by LuaCASC, necessitating an update for this tool to continue
working.

LuaDBC has received an update as well to properly support the WDC4
database type, so that is included. With this the hack of substituting
the header of database files can be removed.