Don't unilaterally override package.path
This is just to make development less of a pain going forward. The
embedded copies of libraries will be picked up now if you cd into the
repository directory and directly execute export.lua, and additionally
we'll now allow the use of any copies of our required libraries that may
be installed elsewhere on the system in preference to our own.