Add support for textual code manifests
In 1.15.4 and 4.4.1 Blizzard appear to be moving toward deprecating and
perhaps removing the manifest database used for exports.

Instead, there's a trio of line-delimited text files shipped in the data
for both of these clients. The 'exportinterfacefiles code' command has
been hooked up to read these instead of the manifest, and exports
whatever is named inside of them.

At present this support is limited to just code exports; art assets for
now at least are continuing to use the old manifest database.

This changeset imbues the export script with the knowledge of the text
manifests and how to process them. If we're explicitly dumping the code
for a product branch, we'll test if the build that we're dumping from
has one of the required txt files and - if so - will read the full list
of files from that instead of consulting the manifest database. If the
file isn't found, then we'll use the old manifest logic.

Note that for now this only works for explicit code exports - dumping
all interface files won't trigger the manifest path for now. Would
rather wait and see what Blizzard do for those before worrying about
that case.